requirements.txt
jsonschema
pyyaml
pytest-factoryboy>=2.0.3
pytest>=5.4.3